Seperation component concern out of index.js.
